// src/app/folha/operadoraplanosaude/models/operadoraplanosaude.ts
export class OperadoraPlanoSaude {
    id : string;
    grupoEmpresaId: string;
    empresaId: string;
    codigo: string;
    sigla: string;
    nome: string;
    cnpj: string;
    registroAns: string;
    telefone: string;
    cep: string;
    tipoLogradouroId: number;
    ufId: number;
    paisId: number;
    localidadeId: number;
    logradouro: string;
    numero : string;
    complemento: string;
    bairro: string;
    codigoMunicipioIbge: string;
    excluido: string;
}